adventofcode v0.1.0 Day1

Functions

calculate_additional_fuel(val, sum \\ 0)

Calculate the fuel required given a mass plus the additional fuel for carrying that fuel

Examples

iex> Day1.calculate_additional_fuel(12)
2

iex> Day1.calculate_additional_fuel(1969)
966

iex> Day1.calculate_additional_fuel(100756)
50346

calculate_fuel(val)

Calculate the fuel required given a mass

Examples

iex> Day1.calculate_fuel(12)
2

iex> Day1.calculate_fuel(14)
2

iex> Day1.calculate_fuel(1969)
654

iex> Day1.calculate_fuel(100756)
33583

solve(input)

Given a list of ints will calculate the fuel for each and sum the result

uses calculate_fuel/1 to calculate fuel

solve2(input)

Given a list of ints will calculate the fuel + additional fuel for each and sum the result

uses calculate_additional_fuel/1 to calculate fuel