adventofcode v0.1.0 Intcode

Intcode seems to be important this year for Advent of Code, so I'm preemptively moving out some helper functions here.

Functions

execute(arg)

Executes the instruction defined by the opcode at ip

Returns a tuple of the new memory and the new instruction pointer. The new instruction pointer is found by incrementing the current instruction pointer by 1 + the number of parameters of this instruction. The instruction pointer can also be set by an instruction, and if it's value is :halt the program halts.

execute(memory, atom)

Executes the given intcode program in memory (a list of integers) and returns the memory after the program halts.

This sets the instruction pointer to 0 initially.

execute_instruction(memory, params, atom)

execute_instruction(memory, list, atom, fun)

get_addr(memory, address)

Gets the value at address in memory

get_parameters_for_instruction(memory, ip, instruction)

Returns a list of all the parameters for the instruction as given by the ip (instruction pointer) and the number of parameters count from the memory

parse_instruction(memory, ip)

Parses the instruction beginning at ip

Returns a tuple of the instruction and it's parameters

update_instruction_pointer(ip, instruction)

update_memory(initial_memory, updates)

Given the initial_memory applies the updates, which are {index, new_value} pairs