adventofcode v0.1.0 Utils

Various Utility functions for solving advent of code problems.

Functions

get_input(day, part, options \\ [])

Reads a file located at inputs/input-{day}-{part}.txt

Options

:stream if true will use File.stream!/1 defaults to File.read!/1

:split if true will split the input line by line

Examples

iex> Utils.get_input(0, 0)
"Test File\nWith Lines"

iex> Utils.get_input(0, 0, split: true)
["Test File", "With Lines"]

iex> Utils.get_input(0, 0, stream: true) |> Stream.run
:ok

md5(value)

Generates the md5 hash of a value and encodes it as a lowercase base16 encoded string.

Examples

iex> Utils.md5("advent of code")
"498fa12185ebe8a9231b9072da43c988"

permutations(list)

Generates all the permutations for the input list

Examples

iex> Utils.permutations([1, 2, 3])
[[1, 2, 3], [1, 3, 2], [2, 1, 3], [2, 3, 1], [3, 1, 2], [3, 2, 1]]

sample(value, threshold \\ 0.999)

Inspects a value, but only if a random value generate is greater than threshold

This is intended to be used with large streams of data that you want to investigate without printing every value.

swap(list, pos_a, pos_a)

Swaps the element at pos_a in list with the element at pos_b

Examples

iex> Utils.swap([1, 2, 3], 0, 1)
[2, 1, 3]

time(fun)

Run the function fun and returns the time in seconds elapsed while running it