adventofcode v0.1.0 Utils.Matrix

A few functions on top of the Matrex library for manipulating matrices.

Functions

apply_to_sub_rect(matrix, x, y, w, h, fun)

Given a matrix will apply the function fun to every element in the sub-matrix starting at [x, y] with width w and height h

shift_col(matrix, x, amount)

Shifts the column x by amount

shift_row(matrix, y, amount)

Shifts the row y by amount

Here we just transpose then shift_col/3 then transpose again